added a new column manager and gave reference as used_id and removed unnecessary methods in models

+++ db/migrate/20150630092013_remove_table_project_employee.rb
@@ -1,9 +1,7 @@
 class RemoveTableProjectEmployee < ActiveRecord::Migration
   def up
-     drop_table :Project_employees
-  end
-  
-  def down
-     raise ActiveRecord::IrreversibleMigration
+    if ActiveRecord::Base.connection.table_exists? 'Project_employees'
+      drop_table :Project_employees
+    end
   end
 end
